requirements.txt
torch
torchvision
pillow
numpy
flask
flask_cors
